+++ app/controllers/all/openers/projects_controller.rb
@@ -1,9 +1,11 @@
-class ProjectsOpenersController < ApplicationController
-  def openers
-    authorize Project
+class All::Openers::ProjectsController < ApplicationController
+  def index
+    authorize Project, :index?
 
     @projects = ProjectsFetcher.new.sorted_openers(month, year)
     @date = "#{Date::MONTHNAMES[month.to_i]} #{year}"
+    @month = month
+    @year = year
   end
 
   private def month