+++ app/controllers/all/revised_conversion_date/projects_controller.rb
@@ -1,4 +1,4 @@
-class All::ConversionDateChanged::ProjectsController < ApplicationController
+class All::RevisedConversionDate::ProjectsController < ApplicationController
   after_action :verify_authorized
   rescue_from ActiveRecord::RecordNotFound, with: :not_found_error
 
@@ -7,6 +7,8 @@ class All::ConversionDateChanged::ProjectsController < ApplicationController
 
     @projects = Project.conversion_date_revised_from(month, year)
     @date = "#{month_name(month)} #{year}"
+    @month = month
+    @year = year
   end
 
   private def month_name(month)